+++ unit-test/openvibe-module-socket/uoSocketClientServerASyncCommunicationTest.cpp
@@ -145,7 +145,7 @@ int uoSocketClientServerASyncCommunicationTest(int argc, char* argv[])
 	{
 		std::string dataString = baseData + std::to_string(sendIndex);
 		std::strcpy(dataBuffer, dataString.c_str());
-		unsigned int dataSize = dataString.size();
+		uint32_t dataSize = static_cast<uint32_t>(dataString.size());
 
 		sendData(client, &dataSize, sizeof(dataSize));
 		sendData(client, dataBuffer, dataSize);
